didi some more enhancement
