+++ pages/registerPage.ts
@@ -1,7 +1,7 @@
 import { Page } from "@playwright/test";
 export default class RegisterPage {
   constructor(public page: Page) {}
-
+// Need to work in future if time permits
   async enterEmail(email: string) {
     await this.page.locator("input[name='email']").type(email);
   }
